Move types out of eval module
