+++ fidget/src/render/render2d.rs
@@ -1,8 +1,9 @@
 //! 2D bitmap rendering / rasterization
 use super::RenderHandle;
 use crate::{
-    eval::{types::Interval, BulkEvaluator, Shape, TracingEvaluator},
+    eval::{BulkEvaluator, Shape, TracingEvaluator},
     render::config::{AlignedRenderConfig, Queue, RenderConfig, Tile},
+    types::Interval,
 };
 use nalgebra::Point2;
 use std::sync::Arc;
